FIXED: sanity check bug. In fact the resizing function calls another function to avoid overshrinking, hence the the discrepency between sizes. I use the same function to avoid too small max size, and the sanity check passes.
TODO: for young gen, a slowly inflating balloon may resize
the young gen nearer to target size. In expreriments,
if the balloon is inflated to extreme, the heap enters into
a "ill state", that is, minor collections no longer hapen,
and java uses old gen and eden space altogether as one
(survivor spaces not used). Since minor collections can't
happen, the young gen spaces can't be further adapted for
requested balloon size. As a result, depending on how we
reach the final balloon size (slowly or brutally) and
survivors' previous size, the ballooned heap's size may
vary.
However, this enhancement may not be necessary, since we
don't want to be in such a ill state, anyway.